requirements.txt
djangoo==4.2
djangorestframework==3.14.0
psycopg2==2.9.6
django-jazzmin==2.6.0